added "Login via Facebook" button, which redirects the user from the Angular frontend to the backend, which redirects then to Facebook. The user accepts and is redirected back to the backend, where the user account is getting extracted. The backend again redirects the user to the frontend, where Angular pops in and fetches the credentials and other information like the name and deletes them from the url. The user may register with the pre-populated fields now and in the back, the access token and the Facebook profile id is sent with it.
When the user comes back after a while and wants to sign in, he has the chance to do so by entering his or her email or by clicking on the "Login via Facebook" button. By doing so, we're redirecting the user to our server, from where he is redirected to Facebook. Facebook sends us the information we previously stored in the Account model. Therefore, we can check for an account with the correct provider ID and if we found it, we may fetch the assigned user and generate a token from that specific user.
The user is redirected back to the frontend, but this time, a JWT is appended to the URL, which angular will capture, read and validate. After successfully validating it, the token is set and the user is redirected to the "users" state we know from this very example.

Simple, isn't it?